driver

+++ Classes/Core/App.mm
@@ -857,7 +857,7 @@ public:
         PASS;
     }
     
-    virtual uint count() const
+    virtual usize count() const
     {
         return argu.size();
     }
